Proof of concept reworking of the autocenter functions to address the discrepancy between the Windows API only allowing toggling autocenter on or off, and the Linux API only allowing setting the strength of the effect.
This leads to an awkward situation for Wine where it's impossible to
enable autocenter without overriding the users preference.

Introduces a new autocenter_strength sysfs entry for setting the
strength but not applying it. Rework the old sysfs autocenter
interface so it toggles the effect on with a non-zero value and off
with zero but does not set the value. This keeps compatibility with
the Linux interface.